dashboard/db_functions.py
import pandas as pd
import altair as alt
import streamlit as st
from datetime import timedelta, datetime

from os import environ

from dotenv import load_dotenv
import sqlalchemy as db
from sqlalchemy.engine.base import Connection


load_dotenv()
STR_DAY_AGO = (datetime.now() - timedelta(hours=24)
               ).strftime("%Y-%m-%d %H:%M:%S")
TODAY = datetime.today()


def get_database_engine():
    """Returns the database engine."""
    try:
        engine = db.create_engine(
            f"mssql+pymssql://{environ['DB_USER']}:{environ['DB_PASSWORD']}@{environ['DB_HOST']}:{environ['DB_PORT']}/{environ['DB_NAME']}?charset=utf8"
        )
        return engine
    except Exception as e:
        print(f"Error creating database engine: {e}")
        raise e


if __name__ == "__main__":

    db_engine = get_database_engine()
