dashboard/requirements.txt
pandas
altair
sqlalchemy
streamlit
pytest
pylint
ipykernel
python-dotenv
pymssql